requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
gunicorn==21.2.0
Werkzeug==2.3.8
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
requests==2.31.0
APScheduler==3.10.4
fpdf2==2.7.9
psycopg2-binary
python-dotenv
beautifulsoup4==4.12.3
sib-api-v3-sdk
cloudinary
Flask-Caching
Flask-Limiter
rich
pypdf==3.17.1
openpyxl

# --- AI & FIXES ---
# New Google AI SDK (Replaces google-generativeai)
google-genai

# CRITICAL FIX: google-api-core 2.28.0+ crashes on Python 3.9
# We pin it to a safe version until you upgrade to Python 3.10+
google-api-core<2.28.0